source code
In this folder, there are source code of construting 256-level BFS Algorithm, accurate query shortest distance in a graph Algorithm, Dijkstra Algorithm and so on. You can judge the function of the Algorithm from reading the name of the file. As well as, you can compile a "*.cpp" with the command "g++ *.cpp" on the platform of which the g++ compiler has been installed.